REFACTOR: pull out interfaces and declar in interfaces/index.ts
