+++ src/Account.ts
@@ -2,7 +2,7 @@
 import { loadStdlib } from '@reach-sh/stdlib';
 // @ts-ignore
 import { masterVault as backend } from '@xbacked-dao/xbacked-contracts';
-import { Vault, UserVaultReturnParams, VaultReturnParams } from './Vault';
+import { Vault } from './Vault';
 import { convertToMicroUnits, calculateInterestAccrued } from './utils';
 import { AccountInterface } from './interfaces';
